+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- IsPublished will no also handle packages that are not published to pub.dev
+
 ## [3.0.3] - 2024-04-09
 
 ### Fixed
@@ -85,6 +91,7 @@
 
 - Initial version.
 
+[Unreleased]: https://github.com/inlavigo/gg_publish/compare/3.0.3...HEAD
 [3.0.3]: https://github.com/inlavigo/gg_publish/compare/3.0.2...3.0.3
 [3.0.2]: https://github.com/inlavigo/gg_publish/compare/3.0.1...3.0.2
 [3.0.1]: https://github.com/inlavigo/gg_publish/compare/3.0.0...3.0.1